Fix pathfinding one more time - just needs to search for a walkable neighbor to dest
The issue was that you could walk over multiple unwalkable neighbor tiles to get to the destination before this commit. The algorithm was slightly off - it was counting each neighbor of the player's walkable island as walkable. Instead, what we needed was to just search for whether the desired tile destination is next to one of the player's walkable island tiles.